Move IntIdTableWithPossibleDependency out of Schema.kt
